[amdgcn] Merge LowerSRegBlockArgs into AMDGCNBufferization
Integrate SREG block-argument lowering directly into handleBlockArgument:
for special registers (SCC, VCC, ...) use an unallocated SGPR slot as the
phi carrier instead of regTy.getAsUnallocated(), which is invalid for those
types. The unallocated SGPR form gives the phi-forward copy a resource
side-effect so it is not eliminated as dead code. Remove LowerSRegBlockArgs
as a standalone pass and update the test accordingly.
