+++ lib/Dialect/AMDGCN/Transforms/CMakeLists.txt
@@ -11,7 +11,6 @@ add_mlir_library(AMDGCNTransforms
   LDSAlloc.cpp
   LegalizeCF.cpp
   LegalizeOperands.cpp
-  LowerSRegBlockArgs.cpp
   LowLevelScheduler.cpp
   Mem2Reg.cpp
   OptimizeAMDGCN.cpp
